Generic OwnerEditor; Create TableOwnerEditor
